refactor: split agent k8s backend lifecycle
