+++ services/api-rs/crates/centaur-sandbox-agent-k8s/src/backend/io.rs
@@ -1,0 +1,48 @@
+use std::pin::Pin;
+
+use centaur_sandbox_core::{
+    SandboxBackend, SandboxError, SandboxId, SandboxIo, SandboxResult, SandboxStatus,
+};
+use kube::api::AttachParams;
+use tokio::io::{AsyncRead, AsyncWrite};
+
+use super::{AgentSandboxBackend, map_kube_error};
+
+impl AgentSandboxBackend {
+    pub(in crate::backend) async fn attach_io(&self, id: &SandboxId) -> SandboxResult<SandboxIo> {
+        if self.status(id).await? != SandboxStatus::Running {
+            return Err(SandboxError::NotReady(format!(
+                "agent sandbox {} is not running",
+                id.as_str()
+            )));
+        }
+        let params = AttachParams::default()
+            .container(self.config.container_name.clone())
+            .stdin(true)
+            .stdout(true)
+            .stderr(true)
+            .tty(false)
+            .max_stdout_buf_size(1024 * 1024)
+            .max_stderr_buf_size(1024 * 1024);
+        let mut attached = self
+            .pods()
+            .attach(id.as_str(), &params)
+            .await
+            .map_err(|err| map_kube_error("attach sandbox pod", err))?;
+        let stdin = attached
+            .stdin()
+            .map(|stream| Box::pin(stream) as Pin<Box<dyn AsyncWrite + Send>>);
+        let stdout = attached
+            .stdout()
+            .map(|stream| Box::pin(stream) as Pin<Box<dyn AsyncRead + Send>>);
+        let stderr = attached
+            .stderr()
+            .map(|stream| Box::pin(stream) as Pin<Box<dyn AsyncRead + Send>>);
+        let stdin = stdin.ok_or_else(|| SandboxError::Io("stdin was not attached".to_owned()))?;
+        let stdout =
+            stdout.ok_or_else(|| SandboxError::Io("stdout was not attached".to_owned()))?;
+        let stderr =
+            stderr.ok_or_else(|| SandboxError::Io("stderr was not attached".to_owned()))?;
+        Ok(SandboxIo::with_guard(stdin, stdout, stderr, attached))
+    }
+}
