+++ services/api-rs/crates/centaur-sandbox-agent-k8s/src/backend/lifecycle.rs
@@ -1,0 +1,130 @@
+use std::time::Duration;
+
+use async_trait::async_trait;
+use centaur_sandbox_core::{
+    ObservedSandbox, SandboxBackend, SandboxError, SandboxHandle, SandboxId, SandboxIo,
+    SandboxResult, SandboxSpec, SandboxStatus,
+};
+use kube::api::{DeleteParams, ListParams, PostParams};
+use tokio::time::{Instant, sleep};
+
+use super::{AgentSandboxBackend, is_not_found, map_kube_error};
+use crate::resources::*;
+use crate::{BACKEND_NAME, MANAGED_BY_LABEL, MANAGED_BY_VALUE};
+
+impl AgentSandboxBackend {
+    async fn wait_until_running(&self, id: &SandboxId) -> SandboxResult<()> {
+        let deadline = Instant::now() + self.config.ready_timeout;
+        loop {
+            match self.status(id).await? {
+                SandboxStatus::Running => return Ok(()),
+                SandboxStatus::Gone | SandboxStatus::Stopped => {
+                    return Err(SandboxError::NotReady(format!(
+                        "sandbox {} reached terminal state before running",
+                        id.as_str()
+                    )));
+                }
+                status if Instant::now() >= deadline => {
+                    return Err(SandboxError::NotReady(format!(
+                        "sandbox {} did not become running before timeout; latest status: {status:?}",
+                        id.as_str()
+                    )));
+                }
+                _ => sleep(Duration::from_millis(500)).await,
+            }
+        }
+    }
+}
+
+#[async_trait]
+impl SandboxBackend for AgentSandboxBackend {
+    fn name(&self) -> &'static str {
+        BACKEND_NAME
+    }
+
+    async fn create(&self, spec: SandboxSpec) -> SandboxResult<SandboxHandle> {
+        let id = SandboxId::new(next_sandbox_name());
+        let resolved_iron_proxy = self.resolve_iron_proxy(&id, &spec)?;
+        if let Err(err) = self
+            .create_iron_proxy_resources(&id, resolved_iron_proxy.as_ref())
+            .await
+        {
+            let _ = self.delete_iron_proxy_resources(&id).await;
+            return Err(err);
+        }
+        let sandbox = build_agent_sandbox(&id, &spec, &self.config, resolved_iron_proxy.as_ref())?;
+        let create_result = self
+            .sandboxes()
+            .create(&PostParams::default(), &sandbox)
+            .await
+            .map_err(|err| map_kube_error("create sandbox", err));
+        if let Err(err) = create_result {
+            let _ = self.delete_iron_proxy_resources(&id).await;
+            return Err(err);
+        }
+        if let Err(err) = self.wait_until_running(&id).await {
+            let _ = self.stop(&id).await;
+            return Err(err);
+        }
+        Ok(SandboxHandle::new(id, BACKEND_NAME))
+    }
+
+    async fn open_io(&self, id: &SandboxId) -> SandboxResult<SandboxIo> {
+        self.attach_io(id).await
+    }
+
+    async fn status(&self, id: &SandboxId) -> SandboxResult<SandboxStatus> {
+        let Some(sandbox) = self.get_sandbox(id).await? else {
+            return Ok(SandboxStatus::Gone);
+        };
+        let replicas = sandbox.spec.replicas.unwrap_or(1);
+        let pod = self.get_pod(id).await?;
+        Ok(sandbox_status_from_pod(replicas, pod.as_ref()))
+    }
+
+    async fn observe(&self, id: &SandboxId) -> SandboxResult<ObservedSandbox> {
+        let status = self.status(id).await?;
+        Ok(ObservedSandbox::new(id.clone(), BACKEND_NAME, status))
+    }
+
+    async fn list_observed(&self) -> SandboxResult<Vec<ObservedSandbox>> {
+        let params =
+            ListParams::default().labels(&format!("{MANAGED_BY_LABEL}={MANAGED_BY_VALUE}"));
+        let sandboxes = self
+            .sandboxes()
+            .list(&params)
+            .await
+            .map_err(|err| map_kube_error("list sandboxes", err))?;
+        let mut observed = Vec::with_capacity(sandboxes.items.len());
+        for sandbox in sandboxes.items {
+            let Some(name) = sandbox.metadata.name else {
+                continue;
+            };
+            let id = SandboxId::new(name);
+            observed.push(self.observe(&id).await?);
+        }
+        Ok(observed)
+    }
+
+    async fn stop(&self, id: &SandboxId) -> SandboxResult<()> {
+        match self
+            .sandboxes()
+            .delete(id.as_str(), &DeleteParams::default())
+            .await
+        {
+            Ok(_) => Ok(()),
+            Err(err) if is_not_found(&err) => Ok(()),
+            Err(err) => Err(map_kube_error("delete sandbox", err)),
+        }?;
+        self.delete_iron_proxy_resources(id).await
+    }
+
+    async fn pause(&self, id: &SandboxId) -> SandboxResult<()> {
+        self.patch_replicas(id, 0).await
+    }
+
+    async fn resume(&self, id: &SandboxId) -> SandboxResult<()> {
+        self.patch_replicas(id, 1).await?;
+        self.wait_until_running(id).await
+    }
+}
